Refactor with FOCUSABLE_ENTITIES
